how bizarre

+++ version1.2.1/src/main/java/hu/unideb/inf/model/User.java
@@ -73,7 +73,7 @@ public class User {
         this.nem = nem;
     }
 
-    public double getMagassag() {
+    public int getMagassag() {
         return magassag;
     }
